requirements.txt
streamlit
groq
matplotlib
python-dotenv
pytesseract
pillow
langchain-community
